Refactor for deletion of utils.misc
