+++ kwave/kmedium.py
@@ -3,7 +3,6 @@ from typing import List
 from warnings import warn
 
 import numpy as np
-import kwave.utils.misc as util
 import kwave.utils.checkutils
