+++ kwave/utils/maputils.py
@@ -1,9 +1,11 @@
 import math
 from math import floor
-from typing import Tuple, Optional
+from typing import Tuple, Optional, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
+import scipy
+
 from kwave.utils.tictoc import TicToc
 
 from kwave.utils.matrixutils import matlab_find, matlab_assign, max_nd
@@ -2343,3 +2345,106 @@ def make_spherical_section(radius, height, width=None, plot_section=False, binar
         raise NotImplementedError
 
     return ss, dist_map
+
+
+def focused_bowl_oneil(radius: float, diameter: float, velocity: float, frequency: float, sound_speed: float,
+                       density: float, axial_positions: Union[np.array, float, list] = None,
+                       lateral_positions: Union[np.array, float, list] = None) -> [float, float]:
+    """
+    focused_bowl_oneil calculates O'Neil's solution (O'Neil, H. Theory of
+    focusing radiators. J. Acoust. Soc. Am., 21(5), 516-526, 1949) for
+    the axial and lateral pressure amplitude generated by a focused bowl
+    transducer when uniformly driven by a continuous wave sinusoid at a
+    given frequency and normal surface velocity.
+
+    The solution is evaluated at the positions along the beam axis given
+    by axial_position (where 0 corresponds to the transducer surface),
+    and lateral positions through the geometric focus given by
+    lateral_position (where 0 corresponds to the beam axis). To return
+    only the axial or lateral pressure, set the either axial_position or
+    lateral_position to [].
+
+    Note, O'Neil's formulae are derived under the assumptions of the
+    Rayleigh integral, which are valid when the transducer diameter is
+    large compared to both the transducer height and the acoustic
+    wavelength.
+
+    Args:
+        radius:
+        diameter:
+        velocity:
+        frequency:
+        sound_speed:
+        density:
+        axial_positions:
+        lateral_positions:
+
+    Example:
+        # define transducer parameters
+        radius = 140e-3  # [m]
+        diameter = 120e-3  # [m]
+        velocity = 100e-3  # [m / s]
+        frequency = 1e6  # [Hz]
+        sound_speed = 1500  # [m / s]
+        density = 1000  # [kg / m ^ 3]
+
+        # define position vectors
+        axial_position = np.arange(0, 250e-3 + 1e-4, 1e-4)  # [m]
+        lateral_position = np.arange(-15e-3, 15e-3 + 1e-4, 1e-4)  # [m]
+
+        # evaluate pressure
+        [p_axial, p_lateral] = focused_bowl_oneil(radius, diameter,
+                                                  velocity, frequency, sound_speed, density,
+                                                  axial_position, lateral_position)
+    Returns:
+         p_axial:           pressure amplitude at the axial_position [Pa]
+         p_lateral:         pressure amplitude at the lateral_position [Pa]
+    """
+    float_eps = np.finfo(float).eps
+
+    def calculate_axial_pressure() -> float:
+        # calculate distances
+        B = np.sqrt((axial_positions - h) ** 2 + (diameter / 2) ** 2)
+        d = B - axial_positions
+        E = 2 / (1 - axial_positions / radius)
+
+        # compute pressure
+        P = E * np.sin(k * d / 2)
+
+        # replace values where axial_position is equal to the radius with limit
+        P[np.abs(axial_positions - radius) < float_eps] = k * h
+
+        # calculate magnitude of the on - axis pressure
+        axial_pressure = density * sound_speed * velocity * np.abs(P)
+        return axial_pressure
+
+    def calculate_lateral_pressure() -> float:
+        # calculate magnitude of the lateral pressure at the geometric focus
+        Z = k * lateral_positions * diameter / (2 * radius)
+        lateral_pressure = 2. * density * sound_speed * velocity * k * h * scipy.special.jv(1, Z) / Z
+
+        # replace origin with limit
+        lateral_pressure[lateral_positions == 0] = density * sound_speed * velocity * k * h
+        return lateral_pressure
+
+    # wave number
+    k = 2 * np.pi * frequency / sound_speed
+
+    # height of rim
+    h = radius - np.sqrt(radius ** 2 - (diameter / 2) ** 2)
+
+    p_axial = None
+    p_lateral = None
+
+    if lateral_positions is not None:
+        p_lateral = calculate_lateral_pressure()
+    if axial_positions is not None:
+        p_axial = calculate_axial_pressure()
+
+    return p_axial, p_lateral
+
+    return [p_axial, p_lateral]
+
+
+def ndgrid(*args):
+    return np.array(np.meshgrid(*args, indexing='ij'))
